refactor: refactor code and structure
