+++ datagrid/src/actions_creator/action_creator.js
@@ -1,0 +1,18 @@
+module.exports = {
+
+  fetchPostsRequest:()=>({
+    type: "FETCH_ALL_ITEMS"
+  }),
+  
+  fetchPostsSuccess: (payload)=> ({
+    type: "FETCH_ALL_ITEMS_SUCCESS",
+    payload
+  }),
+  
+ fetchPostsError:()=> {
+    return {
+      type: "FETCH_ALL_ITEMS_ERROR"
+    }
+  }
+};
+
